Clean up CSS styles: Add Gym Location View

+++ VelocityCoders.FitnessPractice.BLL/GymManager.cs
@@ -11,7 +11,25 @@ namespace VelocityCoders.FitnessPractice.BLL
         #region ||=======  GET ITEM | BY GYM-ID  =======||
         public static Gym GetItem(int gymId)
         {
-            return GymDAL.GetItem(gymId);
+            BrokenRuleCollection saveBrokenRules = new BrokenRuleCollection();
+
+            if (gymId <= 0)
+            {
+                saveBrokenRules.Add("Gym", "Invalid ID: " + gymId.ToString());
+                throw new BLLException("There was an error retrieving Gym.", saveBrokenRules);
+            }
+
+            Gym item = GymDAL.GetItem(gymId);
+
+            if (item == null)
+            {
+                saveBrokenRules.Add("Gym", "Could not retrieve record with ID: " + gymId.ToString());
+                throw new BLLException("Error: No record found", saveBrokenRules);
+            }
+            else
+            {
+                return item;
+            }
         }
         #endregion
         #endregion
@@ -20,25 +38,44 @@ namespace VelocityCoders.FitnessPractice.BLL
         #region ||=======  GET COLLECTION | ALL =======||
         public static GymCollection GetCollection()
         {
-            return GymDAL.GetCollection();
+            GymCollection collection = GymDAL.GetCollection();
+            return collection;
         }
         #endregion
         #endregion
 
         #region SECTION 3 ||=======  INSERT OR UPDATE ITEM  =======||
         #region ||=======  INSERT OR UPDATE ITEM | BY GYM-ID  =======||
-        public static int SaveItem(int gymId, Gym gymObj)
+        public static int SaveItem(Gym gymObj)
         {
-            return GymDAL.SaveItem(gymId, gymObj);
+            BrokenRuleCollection saveBrokenRules = new BrokenRuleCollection();
+
+            if (string.IsNullOrEmpty(gymObj.GymName))
+            {
+                saveBrokenRules.Add("Gym Name", "Name is required");
+            }
+
+            if (saveBrokenRules.Count > 0)
+            {
+                throw new BLLException("There was an error saving Gym.", saveBrokenRules);
+            }
+            else
+            {
+                return GymDAL.SaveItem(gymObj);
+            }
+
         }
         #endregion
         #endregion
 
         #region SECTION 4 ||=======  DELETE ITEM  =======||
         #region ||=======  DELETE ITEM | BY GYM-ID  =======||
-        public static int DeleteItem(int gymId)
+        public static bool DeleteItem(int gymId)
         {
-            return GymDAL.DeleteItem(gymId);
+            if (gymId > 0)
+                return GymDAL.DeleteItem(gymId);
+            else
+                throw new BLLException("Delete failed. Gym ID is invalid: " + gymId.ToString());
         }
         #endregion
         #endregion
